+++ flashcards/apps/api/v1/anki.py
@@ -3,7 +3,7 @@ Implementation of AnkiConnect, which works super well :)
 """
 import json
 import urllib.request
-from flashcards.utils import get_csv_from_openai
+from flashcards.apps.api.v1.openai import get_csv_from_openai  # pylint: disable=import-error,no-name-in-module
 
 def request(action, **params):
     """
@@ -30,15 +30,9 @@ def invoke(action, **params):
 
 
 def create_anki_cards(openai_data):
-    result = invoke('deckNames')
-    print('got list of decks: {}'.format(result))
-
-    # csv = """What is food fermentation?,A biochemical process using microorganisms to produce diverse foods
-    # What are some examples of popular fermented foods?,Beer, yogurt, pickles
-    # What foods can you create with microbial environments?,Mead, sourdough, tempeh, and more
-    # What is the purpose of this course?,To explore the role of microbes in food fermentation
-    # What will you study in this course?,Different types of fermentations"""
-
+    """
+    Generates anki cards from data
+    """
     rows = openai_data.split('\n')
     for row in rows:
         question, answer = row.split(',', 1)
@@ -53,25 +47,22 @@ def create_anki_cards(openai_data):
                 #     "yomichan"
                 # ],
             }
-        print('\nAdding card:', note)
         invoke('addNote', note=note)
 
 def main():
+    """
+    Master function that gets a response from openai and passes the result to Anki
+    """
     result = get_csv_from_openai()
     # TODO: Insert some kind of data validation here to make sure openai sent back something nice
-
     result = result.replace('\t','')
-    # print(result)
     create_anki_cards(result)
 
 main()
-# invoke('createDeck', deck='test1')
 
 
 
-
-
-# Below is some code using just the plain ol' anki package, which failed to work on my machine.
+# Below is some code using just the plain ol' anki package, which has failed to work on my machine so far.
 
 # import anki
 # # from anki.collection import ImportCsvRequest
